// lib/types.ts
export interface DisasterData {
  id: string
  no: number
  kecamatan: string
  jumlah_penduduk: number
  meninggal: number
  luka: number
  hilang: number
  pengungsi_di_luar_pandan: number
  terdampak: number
  rumah_rusak_ringan: number
  rumah_rusak_sedang: number
  rumah_rusak_berat: number
  sekolah_rusak_ringan: number
  sekolah_rusak_sedang: number
  sekolah_rusak_berat: number
  created_at: string
  updated_at: string
}